+++ glacier2/src/ModelViewerState.cpp
@@ -58,7 +58,7 @@ namespace Glacier {
       gEngine->getActionManager()->getCameraController()->getMovement()
     );
 
-    mDirector->getCamera()->update( tick );
+    mDirector->getCamera()->update( (float)tick );
   }
 
   void ModelViewerState::shutdown( GameTime time )
